add pre-trained model to HiNAS
